+++ changelog.md
@@ -55,6 +55,7 @@ __                                                                              
 * Alternative Shed Ceiling
 * Hunt House on tundra map
 * Increse critical chance if you dont walk from 15% to 25%
+* Bounce Kill Score Reward
 __                                                                                                                 __
 ## Changes
 __                                                                                                                 __